ensure album pages are static
